Extracting abstract client from bft pyclient
As part of this commit, methods from the bft python client BftClient are
extracted into an abstract class called Client. These classes are
located in bft_client.py file. The aim is to ensure that the
ClientService class in the
concord-bft/tests/apollo/clientservice/csutil.py file can extend this
abstract Client class. This would enable the ClientService to be used in
the ApolloTests.